Add option for custom routes handling
