+++ packages/next-auth/src/core/index.ts
@@ -1,14 +1,12 @@
 import logger, { setLogger } from "../lib/logger"
-import * as routes from "./routes"
 import renderPage from "./pages"
 import { init } from "./init"
 import { assertConfig } from "./lib/assert"
 import { SessionStore } from "./lib/cookie"
 
-import type { NextAuthOptions } from "./types"
+import { NextAuthOptions } from "./types"
 import type { NextAuthAction } from "../lib/types"
 import type { Cookie } from "./lib/cookie"
-import type { ErrorType } from "./pages/error"
 
 export interface IncomingRequest {
   /** @default "http://localhost:3000" */
@@ -86,151 +84,12 @@ export async function NextAuthHandler<
     options.logger
   )
 
-  if (method === "GET") {
-    const render = renderPage({ ...options, query: req.query, cookies })
-    const { pages } = options
-    switch (action) {
-      case "providers":
-        return (await routes.providers(options.providers)) as any
-      case "session": {
-        const session = await routes.session({ options, sessionStore })
-        if (session.cookies) cookies.push(...session.cookies)
-        return { ...session, cookies } as any
-      }
-      case "csrf":
-        return {
-          headers: [{ key: "Content-Type", value: "application/json" }],
-          body: { csrfToken: options.csrfToken } as any,
-          cookies,
-        }
-      case "signin":
-        if (pages.signIn) {
-          let signinUrl = `${pages.signIn}${
-            pages.signIn.includes("?") ? "&" : "?"
-          }callbackUrl=${options.callbackUrl}`
-          if (error) signinUrl = `${signinUrl}&error=${error}`
-          return { redirect: signinUrl, cookies }
-        }
-
-        return render.signin()
-      case "signout":
-        if (pages.signOut) return { redirect: pages.signOut, cookies }
-
-        return render.signout()
-      case "callback":
-        if (options.provider) {
-          const callback = await routes.callback({
-            body: req.body,
-            query: req.query,
-            headers: req.headers,
-            cookies: req.cookies,
-            method,
-            options,
-            sessionStore,
-          })
-          if (callback.cookies) cookies.push(...callback.cookies)
-          return { ...callback, cookies }
-        }
-        break
-      case "verify-request":
-        if (pages.verifyRequest) {
-          return { redirect: pages.verifyRequest, cookies }
-        }
-        return render.verifyRequest()
-      case "error":
-        // These error messages are displayed in line on the sign in page
-        if (
-          [
-            "Signin",
-            "OAuthSignin",
-            "OAuthCallback",
-            "OAuthCreateAccount",
-            "EmailCreateAccount",
-            "Callback",
-            "OAuthAccountNotLinked",
-            "EmailSignin",
-            "CredentialsSignin",
-            "SessionRequired",
-          ].includes(error as string)
-        ) {
-          return { redirect: `${options.url}/signin?error=${error}`, cookies }
-        }
-
-        if (pages.error) {
-          return {
-            redirect: `${pages.error}${
-              pages.error.includes("?") ? "&" : "?"
-            }error=${error}`,
-            cookies,
-          }
-        }
-
-        return render.error({ error: error as ErrorType })
-      default:
-    }
-  } else if (method === "POST") {
-    switch (action) {
-      case "signin":
-        // Verified CSRF Token required for all sign in routes
-        if (options.csrfTokenVerified && options.provider) {
-          const signin = await routes.signin({
-            query: req.query,
-            body: req.body,
-            options,
-          })
-          if (signin.cookies) cookies.push(...signin.cookies)
-          return { ...signin, cookies }
-        }
-
-        return { redirect: `${options.url}/signin?csrf=true`, cookies }
-      case "signout":
-        // Verified CSRF Token required for signout
-        if (options.csrfTokenVerified) {
-          const signout = await routes.signout({ options, sessionStore })
-          if (signout.cookies) cookies.push(...signout.cookies)
-          return { ...signout, cookies }
-        }
-        return { redirect: `${options.url}/signout?csrf=true`, cookies }
-      case "callback":
-        if (options.provider) {
-          // Verified CSRF Token required for credentials providers only
-          if (
-            options.provider.type === "credentials" &&
-            !options.csrfTokenVerified
-          ) {
-            return { redirect: `${options.url}/signin?csrf=true`, cookies }
-          }
-
-          const callback = await routes.callback({
-            body: req.body,
-            query: req.query,
-            headers: req.headers,
-            cookies: req.cookies,
-            method,
-            options,
-            sessionStore,
-          })
-          if (callback.cookies) cookies.push(...callback.cookies)
-          return { ...callback, cookies }
-        }
-        break
-      case "_log":
-        if (userOptions.logger) {
-          try {
-            const { code, level, ...metadata } = req.body ?? {}
-            logger[level](code, metadata)
-          } catch (error) {
-            // If logging itself failed...
-            logger.error("LOGGER_ERROR", error as Error)
-          }
-        }
-        return {}
-      default:
-    }
-  }
-
-  return {
-    status: 400,
-    body: `Error: Action ${action} with HTTP ${method} is not supported by NextAuth.js` as any,
-  }
+  return options.routesHandler({
+    options,
+    error,
+    cookies,
+    sessionStore,
+    req,
+    userOptions,
+  }) as OutgoingResponse<Body>
 }
