+++ packages/next-auth/src/index.ts
@@ -2,5 +2,6 @@ export * from "./core/types"
 
 export type { IncomingRequest, OutgoingResponse } from "./core"
 
+export * from "./core/handle-routes"
 export * from "./next"
 export { default } from "./next"
